requirements.txt
pandas
selenium
webdriver-manager
openpyxl
pytesseract 
pillow
numpy
opencv-python
scipy
# img_src = "data:image/png;base64,R0lGODdhNgFQAIQAAHRydLy6vNze3JSWlPTy9KyqrMzOzISGhOzq7KSipMTGxPz6/LSytNTW1IyOjOTm5JyenHR2dLy+vOTi5JyanPT29KyurNTS1IyKjOzu7KSmpMzKzPz+/LS2tNza3JSSlCwAAAAANgFQAAAF/iAnjmQJAFGprmzrvnAsu9Fp3+is73zvw6hb6kcsnjDFJC0yVBIxOBvSSa1aOUwO9FS5elc1wFdpG/MO0aN5zSZFPqJbu2ktm9qvsDhZu/cqaQBTeDJ7hC6GHDYHh1dbdIqJjVhyVXuSZ2mQk5wwWSMONl2dTls5PZhFoaKkMICBg60qqUm0Im8mdiu2PqFjOLGTqyeMJLydFWhpwbKcqYstxz2bdTfMeDiz0siBgs1sJ27Uw6PfiNQsyt6NOOUs4bWoK6+B6OY0O7i7ukCXu2bbcp3Co2ebr3g8bCXrJmWWPT4OjEWk4svWsIAx4C0htEUHxhUX3d3jsZDhspGI/iYaQ8TKy0czYWbYOnaxGEoiFeiZ3MnzxEMZB2eBfKjHJpifiO5p9CFmqTZiNz266FKyp1WeZhwQvSFSSQSV5mL+oNkuqpecHNBWvcrWVp+MJSJci1Swq9eoTj2tbHHx5UuzKtRyUUu1cFrDPJHKBPsUKmAj2n5x1VFD8eN3YuzKEMxwLtCfBY1eVsSYh9hqLaWO9nG6SE6dOCyX0Pqi7GpQMzPm5VPp9urdadVF8ayCNktovi0BZz2ZCoDS5v72wuCguvXrB4QLGfHh+nUsxCmlTj5iOYwt1KXHbY6FjPrLFOJTkFDiA/Xq1PNjgLLfwf0DDmQHixb7FZjeV9+V/hfHeOQp+IMp74lQU3nm7YMSHCUg4F118TUAw4bVzaaffySukp8g+AUoICyr+HeEgddNwZ4ivzxW0A8TGuNXhGOJASIz/hmYH4cvBAmjiCQOCSMGNfR3X3UrLpNegdaNiF9vYxUiWz48LsgPX4bk2CAJRlIJ0pNJkuhCiVUyluaQbSqppggDiLAWDgAmCWd/M8KlozN+jPBJQjb8lIhtZhWQYAkVsIkfC3q+GSKk+g2JpKVFaIelpBB66BF0QFTzRV/hcdDXGhV+aGYJCyi535pyukppiZiCouQCSsC2nQoS2LeIaF+kqlQxX+7DIA9BbSIsrI+SEICTc/IVa5qz/sZ6qX9UaBqbCurk6eCYkvGCKI5dFplfhtBSpyqn0ZI5rbruJulcBNlpu+sNAMILrnOY0ZLNSPaUSIIH1q77LnTs6mtru4TeIleAKtqbRr6gbrnvt5E1BoBmUckLyptAHaywhEMSTO21XnHKSE5tsWXxxWq40edlApN8cgwJM2yqfhwMoGRxQ3JciM1oqkRVy1ihUu4hHeF27EhOebyz1ElNffC1cKRZmqQgdj2pqU28eW5gSDN0ccMS4ntZEPFOFEqtG1nNLspyAxmrkHgfCXa8coJ0Qwb11vva4IQTftuy/VyCXHJ6csAAf9gOXXdEN3Mwwc9ybz2t16GkuDff/lo7QEF3puJ7AAYAeksQGcFGI0ZdRDCwoQU+WNB1AkCrSWWpZ6b58NhEqwTyta7mLQV1cuV+N4kFAZBnm4SU+5xuTqQhtAsT4Ie3dTokUCbeGC6859c6SCr3CBTcrPWgk8Pp3X7Ou3hNzunhma+TChCCDvv5gEpEIhByXi8cFSfqIAAG3Sna+9wEOVf5T1oeQ5G+Kvcm/iVMSNU5QpOqxakoZCdF/skAIYJSBUjwTxV4mo6edqenA0KwUhgcmfy4NpGAaE0RmFsABWWFNQktEH5HEGHvptW8+PXtbNhIoQ40RMN3tSCBoYsT+egnI4Ng7mRp8gDoyBe8UiWPWbFq/l69OKUXJBKhL4WyEPEiB7qRiYAAnJoA32JBPy6aq2ZY5Jn4ggQWH2EOUl8xGLXwVQ6uyREfMoleGy6SgbVAImpvogAHMRC+PQJvj23klCRD5jFqpakAmaSjBHUGCsuw619rlMc/zEghx9hJD3eEWwmyB6cSpO+IswAepy5gPk5O8GqZZMwo3XgL3nXRP7AbIilZecZxtfKFy7TkgiZHn13cx5Juk9MDQYdNWZJxNqMsjj58maSZtZGZynmalya5Ta1t8pjVGsTw2rfN4HWTjdgsTQMZ8wZclS9Wi+OgAzbQikCiRG0s2QTBNvejXnJAArGqpg9xeUMRUGBzgryn/ubmuUUJfXEHktLD9SyAS4yt4WWoKsp58iKy7U1rBLRs4jfPF8opvEV5wnspTq+508/1QFJYgiYzlhYDlALEnNoIRs5+6NAAXHBu9sykGlqDyWMSc6ZbXIcK9RNQaNazRihRaQyadk8Y5mwEJGUXuw5JU5i66ggPJFEiOEWAnZIySRs8I5xEqoJNcJQdN0GqGstKq1MOYgMio9Yo9pCkSnaRGKKTFhRCqUxZ2gyuCvzqYxmUF60RFJ094sItqEdZmQwCATrF2SXbmCcM4A4kQcqkY62KsCaJTRCZ5R7xuhqHnRqTtE646RouwqW5yum1W40mTlnQqhJhdlZgcega/on5IqDCEIO7VacKElBSyPRIKG0QrOsGAUXL/rNysfTf5YIEVzuWTmHSDWVpcMspPmVWhkoSLwliutqqBYuohBIgD9CDzX44oK62fJMBdvEVBlSVBUq6nx0hR4JbkuizdlVYFtTqohcVCK7XQig554uZjO1Lv65bh6SQ6zcUOYB2mVTBAnDLTWbFiGH0ddYfM6ySS1jYgUB0MXqUeiIUtxHDveVFEwDciaAWV8hp8qcKSFqDfNmVgfhE75lGlJ8L8I1uWIAEAwqgATIXAJRhFtTvtJYfZVCsPxBAEg40u98hLRi8XzghS/J8rI8EMFZanA0fddbcI0KURAiOqgt0/ojRhTnYXWfeS5E+SiNK5C5Az+muCAzgTNCGt1g+0I4Db3yf/hAAE4/j2o6t9k4XpDVWUiZzCT0j3P912tNZ0S5TjrfCd8EhFXVUnlm0Mtsr9EWIMmMyoD4tWhQKhGJQ1nQJvOfrEhAMzVfoknH8tOunUXVM0nnJMESD0rxIYRjTuvMLDJBaSbskQpQOVbebjev/7WAyBxlnLYxoxCRhewYTIAAxewuQXug717rec72VJuD2ZDscGdBAktS9cHEewsglNqOeK20evqJK2WOK9zuMCk5XAgbkHmmKuED9cGBVXJzFXuW+6W0Wlh+02QGRDllNjOttX9zkFKFzP5pM/hld5DzcM0H5N0Q+h4TvWinq+YiTxyAWxJ3tDVavHss3DrXBEty7x7leNWT0coIPJHo0Xxue/8QadBwDlmBNdtnTRqFGDGPuBJm6TP5i82XLwDJM/3na8e4FvmbdpLVJgTWMbRrZbOPgzgA64Qvf9483xL91iFtClu70yRuh8vL+gXB+y69sN9kQN/L8UTf2DeGczed230Myx0TCZty99KIvQ4T8cQjYd0JMJbz6vMXODrYhBA++nwYKeRvcq8e8NpJvxs5vE/h7P7LozFc9vzo/iW/b6Pn3NoFsmkd8paxN71+P/Dtk4ZTqHyI0F9/3mpS+98NnKeVMov/aJWd7GxFjA2eBkhGk91+st3+dMH1h4W7sl32BhQchAAA7"
